bazar_frontend_server/requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.5
django-admin==2.0.1
django-excel-base==1.0.4
django-excel-response2==3.0.2
django-filter==2.4.0
django-heroku==0.3.1
django-six==1.0.4
djangorestframework==3.12.4
gunicorn==20.1.0
install==1.3.4
Markdown==3.3.4
psycopg2==2.9.1
psycopg2-binary==2.9.1
pytz==2021.1
screen==1.0.1
sqlparse==0.4.1
whitenoise==5.2.0
xlwt==1.3.0
requests
pymemcache